Collapse lib shim into components and move DQ integration
