+++ src/dc43/components/contract_store/__init__.py
@@ -1,13 +1,13 @@
 """Contract store interfaces, gateways, and reference implementations."""
 
 from dc43.services.contracts.backend.stores.interface import ContractStore
-from .impl.filesystem import FSContractStore
-from .impl.delta import DeltaContractStore
-from .impl.collibra import (
+from .impl import (
     CollibraContractAdapter,
     CollibraContractGateway,
     CollibraContractStore,
     ContractSummary,
+    DeltaContractStore,
+    FSContractStore,
     HttpCollibraContractAdapter,
     HttpCollibraContractGateway,
     StubCollibraContractAdapter,
