+++ src/dc43/integration/dlt_helpers.py
@@ -8,7 +8,7 @@ Translate ODCS DataQuality rules to DLT expectations.
 from typing import Dict
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.lib.components.data_quality.integration import (
+from dc43.integration.data_quality import (
     expectations_from_contract as _expectations_from_contract,
 )
