+++ src/dc43/integration/violation_strategy.py
@@ -12,7 +12,7 @@ except Exception:  # pragma: no cover
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.lib.components.contract_validation import ValidationResult
+from dc43.components.contract_validation import ValidationResult
 
 
 def _merge_pipeline_context(
