--- src/dc43/lib/components/data_quality/validation.py
@@ -1,6 +1,0 @@
-"""Compatibility wrapper for contract-application helpers."""
-
-from dc43.services.data_quality.backend.integration import SPARK_TYPES, spark_type_name
-from dc43.services.data_quality.backend.validation import apply_contract
-
-__all__ = ["apply_contract", "SPARK_TYPES", "spark_type_name"]
