--- src/dc43/lib/data_quality.py
@@ -1,5 +1,0 @@
-"""Compatibility exports for data-quality observation payloads."""
-
-from dc43.services.data_quality.models import ObservationPayload
-
-__all__ = ["ObservationPayload"]
