+++ src/dc43/services/data_quality/backend/integration/__init__.py
@@ -1,29 +1,7 @@
-"""Execution-engine integration helpers for data-quality observations."""
+"""Compatibility exports for engine integrations inside the service backend."""
 
-from .spark import (
-    SPARK_TYPES,
-    attach_failed_expectations,
-    build_metrics_payload,
-    collect_observations,
-    compute_metrics,
-    evaluate_observations,
-    expectations_from_contract,
-    odcs_type_name_from_spark,
-    schema_snapshot,
-    spark_type_name,
-    validate_dataframe,
-)
+from dc43.integration.data_quality import *  # noqa: F401,F403
 
-__all__ = [
-    "SPARK_TYPES",
-    "spark_type_name",
-    "odcs_type_name_from_spark",
-    "schema_snapshot",
-    "expectations_from_contract",
-    "compute_metrics",
-    "collect_observations",
-    "evaluate_observations",
-    "validate_dataframe",
-    "build_metrics_payload",
-    "attach_failed_expectations",
-]
+from dc43.integration import data_quality as _integration
+
+__all__ = list(_integration.__all__)
