+++ tests/test_contract_drafter.py
@@ -11,8 +11,8 @@ from open_data_contract_standard.model import (  # type: ignore
     SchemaProperty,
 )
 
-from dc43.lib.components.contract_drafter import draft_from_validation_result
-from dc43.lib.components.contract_validation import ValidationResult
+from dc43.components.contract_drafter import draft_from_validation_result
+from dc43.components.contract_validation import ValidationResult
 
 
 def _build_contract() -> OpenDataContractStandard:
